requirements.txt
sqlalchemy
aiohttp==1.0.5
aiopixiv==1.0.4
aioredis==0.2.9
async-timeout==1.1.0
asyncio-extras==1.2.0
cffi==1.8.3
chardet==2.3.0
cycler==0.10.0
dateparser==0.5.0
dill==0.2.5
discord.py==1.0.0a0
elasticsearch==5.0.1
elasticsearch-dsl==5.0.0
forbiddenfruit==0.1.2
gitdb2==2.0.0
GitPython==2.1.1
hiredis==0.2.0
jdatetime==1.8.1
Jinja2==2.9.5
Logbook==1.0.0
MarkupSafe==0.23
matplotlib==2.0.0
multidict==2.1.2
multiprocessing-on-dill==3.5.0a4
numpy==1.12.0
objgraph==3.0.1
parsedatetime==2.1
PixivPy==3.2.0
psutil==4.4.2
pycparser==2.17
PyNaCl==1.0.1
pyparsing==2.1.10
python-dateutil==2.5.3
pytz==2016.7
regex==2016.10.22
requests==2.11.1
ruamel.yaml==0.12.15
seaborn==0.7.1
six==1.10.0
smmap2==2.0.1
tabulate==0.7.5
tzlocal==1.3
umalqurra==0.2
urllib3==1.19
uvloop==0.5.4
websockets==3.2
yarl==0.5.3
youtube-dl==2016.11.4
